Migrate createWorkout TS logic into own hook, update imports
