+++ frontend/kettlepal-fe/src/Components/NewWorkouts/FormComponents.tsx/Exercise/ExerciseSets.tsx
@@ -1,6 +1,6 @@
 import { FormControl, FormLabel, Input } from "@chakra-ui/react";
-import { CreateWorkoutState } from "../../CreateWorkout";
 import theme from "../../../../Constants/theme";
+import { CreateWorkoutState } from "../../../../Hooks/useCreateWorkoutForm";
 
 interface ExerciseSetsProps {
   submitted: boolean;
