+++ frontend/kettlepal-fe/src/Components/ViewWorkouts/ViewDetailedWorkoutModal/EditWorkout.tsx
@@ -2,8 +2,6 @@ import {
   HStack,
   VStack,
   Box,
-  Text,
-  useDisclosure,
   Alert,
   AlertIcon,
   AlertDescription,
